wip: move mongodb to framework
