+++ src/framework/mongodb/client/client.ts
@@ -1,5 +1,5 @@
 import { mongodbConfig } from '@/config/mongodb';
-import { createClient } from '@/mongodb/client/create-client';
+import { createClient } from '@/framework/mongodb/client/create-client';
 
 export const mongoDBClient = createClient({
   uri: mongodbConfig.connectionString,
