+++ src/framework/mongodb/schema/workspaces-collection-schema.ts
