+++ src/kairos/authentication/infrastructure/repository/find-one-by-email.test.ts
@@ -1,7 +1,7 @@
 import { invalidCredentialsError } from '@/kairos/authentication/domain/errors';
 import { createFindOneByEmail } from '@/kairos/authentication/infrastructure/repository/find-one-by-email';
 
-import { usersCollection } from '@/mongodb/collection/users-collection';
+import { usersCollection } from '@/framework/mongodb/collection/users-collection';
 import { integrationTest } from '@tests/__vitest__/integration-test';
 import { describe, expect, it } from 'vitest';
