+++ src/kairos/authentication/infrastructure/repository/find-one-by-email.ts
@@ -1,6 +1,6 @@
 import { invalidCredentialsError } from '@/kairos/authentication/domain/errors';
 import { type FindOneByEmail } from '@/kairos/authentication/domain/user-credentials-repository';
-import { type UsersCollection } from '@/mongodb/schema/users-collection-schema';
+import { type UsersCollection } from '@/framework/mongodb/schema/users-collection-schema';
 import { err, ok } from '@/kairos/shared/kernel/result';
 
 type CreateFindOneByEmailDependencies = Readonly<{
