--- src/mongodb/collection/users-collection.ts
@@ -1,4 +1,0 @@
-import { mongoDBClient } from '@/mongodb/client/client';
-import { UsersCollection, usersCollectionName } from '@/mongodb/schema/users-collection-schema';
-
-export const usersCollection: UsersCollection = mongoDBClient.db().collection(usersCollectionName);
