--- src/mongodb/collection/workspaces-collection.ts
@@ -1,4 +1,0 @@
-import { mongoDBClient } from '@/mongodb/client/client';
-import { WorkspacesCollection, workspacesCollectionName } from '@/mongodb/schema/workspaces-collection-schema';
-
-export const workspacesCollection: WorkspacesCollection = mongoDBClient.db().collection(workspacesCollectionName);
