+++ tests/workspace/create-workspace.test.ts
@@ -1,6 +1,6 @@
 import { appConfig } from '@/config';
 
-import { workspacesCollection } from '@/mongodb/collection/workspaces-collection';
+import { workspacesCollection } from '@/framework/mongodb/collection/workspaces-collection';
 import { createTestAgent } from '@koala-ts/framework';
 import { integrationTest } from '@tests/__vitest__/integration-test';
 import { describe, expect, test } from 'vitest';
